refactor: :recycle: seperated schemas in new file
